+++ lib/local_album.g.dart
@@ -14,6 +14,7 @@ LocalAlbum _$LocalAlbumFromJson(Map<String, dynamic> json) {
         : LocalImage.fromJson(json['coverImg'] as Map<String, dynamic>),
     json['title'] as String,
     json['imageCount'] as int,
+    json['videoCount'] as int,
     json['transferType'] as int,
   );
 }
@@ -23,6 +24,7 @@ Map<String, dynamic> _$LocalAlbumToJson(LocalAlbum instance) =>
       'id': instance.id,
       'title': instance.title,
       'imageCount': instance.imageCount,
+      'videoCount': instance.videoCount,
       'coverImg': instance.coverImg?.toJson(),
       'transferType': instance.transferType,
     };