+++ CHANGELOG.md
@@ -4,15 +4,18 @@
 
 ### Breaking
   * Upgraded to New Swift 1.12 plugin structure, may work with older Flutter version but not guaranteed
+  * This version only works with Flutter 1.17.0 and above due to interaction with the Flutter cache
   
 ### New
 * new `getVideoFile` method on `LocalImageProvider` retrieves a file path for a video. 
 * `LocalImage` now includes both images and videos
 * added the `mediaType`, `isVideo` and `isImage` properties to `LocalImage`
 * added `fileName` and `fileSize` properties to `LocalImage`
-* control cache usage with the new `maxCacheDimension` property on `LocalImageProvider`
+* control cache usage with the new `maxCacheDimension` property on `LocalImageProvider`, requires Flutter 1.17.0
 * test coverage improvements
 * pubspec changes for Flutter 1.12, minimum version now 1.12
+* added `videoCount` to `LocalAlbum` 
+* `findImagesInAlbum` now returns both images and video, use the `isVideo` and `isVide` properties to filter
 
 ## 1.0.0